marble/workspace/test_solution.py
Test Implementation:
```python
import pytest
from merge_sort import merge_sort

# Positive Test Cases
def test_merge_sort_positive():
    assert merge_sort([3, 2, 5, 1, 4]) == [1, 2, 3, 4, 5]
    assert merge_sort([-5, -2, -1, -4, -3]) == [-5, -4, -3, -2, -1]

# Edge Cases
def test_merge_sort_edge():
    assert merge_sort([]) == []
    assert merge_sort([1]) == [1]
    assert merge_sort([1000000, 100000, 10000]) == [10000, 100000, 1000000]
    assert merge_sort([5, 5, 3, 3, 1, 1]) == [1, 1, 3, 3, 5, 5]

# Performance Test Cases
def test_merge_sort_performance():
    assert merge_sort(list(range(1000000))) == list(range(1000000))
    assert merge_sort(list(range(1000000, 0, -1))) == list(range(1, 1000001))

# Error Handling Cases
def test_merge_sort_error():
    with pytest.raises(TypeError):
        merge_sort("not an array")
        merge_sort([1, 2, "not an integer"])
        merge_sort(None)
```
In the above test cases, `pytest` is used as the testing framework. Each test function follows the `test_` naming convention that `pytest` uses to automatically identify test cases. The `assert` statement is used to verify the results of the function calls. For error handling cases, the `pytest.raises` context manager is used to check that the function raises the expected exceptions when given invalid inputs.